+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 
 ### Added
 - Function documentation | Mocker Class
+- Support complex column data types
 ### Changed
 - Arrange Mocker generators functions alphabetically | Mocker Class
 - Naming conventions | Mocker Class